+++ lib/src/refs/stream.dart
@@ -51,7 +51,4 @@ class StreamRefState<T> extends AsyncRefState<T, StreamRef<T>> {
 
   @override
   AsyncSnapshot<T> bind() => snapshot;
-
-  @override
-  T read() => snapshot.data as T;
 }